# docs/swagger.json
{
    "openapi": "3.1.0",
    "info": {
      "version": "1.0.0",
      "title": "Contacts app",
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      },
      "description": "This is a documentation of contacts app"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      },
      {
        "url": "https://swagger-contacts-l6oi.onrender.com"
      }
    ],
...
    "paths": {
      "/auth/register": {
        "post": {
          "tags": [
            "Auth"
          ],
          "summary": "Register a new user",
          "operationId": "registerUser",
          "description": "Create new user",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "description": "User registration data",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "description": "User's name",
                      "example": "John Doe"
                    },
                    "email": {
                      "type": "string",
                      "description": "User's email address",
                      "format": "email",
                      "uniqueItems": true,
                      "example": "john.doe@example.com"
                    },
                    "password": {
                      "type": "string",
                      "description": "User's password",
                      "example": "password123"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Successfully registered a user",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/user"
                  }
                }
              }
            },
            "400": {
              "description": "Bad request"
            },
            "409": {
              "description": "Email in use"
            }
          }
        }
      },
      "/auth/login": {
        "post": {
          "tags": [
            "Auth"
          ],
          "summary": "Login a user",
          "operationId": "loginUser",
          "description": "Authenticate user and provide session tokens",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "description": "User login credentials",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "password": {
                      "type": "string",
                      "format": "password"
                    }
                  },
                  "required": [
                    "email",
                    "password"
                  ]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successfully logged in a user",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "accessToken": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "accessToken"
                    ]
                  }
                }
              }
            },
            "400": {
              "description": "Bad request"
            },
            "401": {
              "description": "Unauthorized - Invalid credentials"
            },
            "404": {
              "description": "User not found"
            }
          }
        }
      },
      "/auth/logout": {
        "post": {
          "tags": [
            "Auth"
          ],
          "summary": "Logout a user",
          "operationId": "logoutUser",
          "description": "Invalidate the user session",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "204": {
              "description": "Successfully logged out"
            },
            "400": {
              "description": "Bad request"
            }
          }
        }
      },
...
      "/auth/reset-pwd": {
        "post": {
          "tags": [
            "Auth"
          ],
          "summary": "Reset user password",
          "operationId": "resetPassword",
          "description": "Reset the user password using a reset token",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "description": "Reset password details",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "password": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "token",
                    "password"
                  ]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Password successfully reset"
            },
            "400": {
              "description": "Bad request"
            },
            "401": {
              "description": "Unauthorized - Invalid token"
            },
            "404": {
              "description": "User not found"
            }
          }
        }
      },
...
    },
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer"
        }
      },
      "schemas": {
        "user": {
          "type": "object",
          "required": [
            "name",
            "email",
            "password"
          ],
          "properties": {
            "name": {
              "type": "string",
              "description": "User's name",
              "example": "John Doe"
            },
            "email": {
              "type": "string",
              "description": "User's email address",
              "format": "email",
              "uniqueItems": true,
              "example": "john.doe@example.com"
            },
            "password": {
              "type": "string",
              "description": "User's password",
              "example": "password123"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time",
              "description": "Timestamp when the user was created",
              "example": "2023-08-02T10:20:30Z"
            },
            "updatedAt": {
              "type": "string",
              "format": "date-time",
              "description": "Timestamp when the user was last updated",
              "example": "2023-08-02T10:20:30Z"
            }
          }
        },
...
      },
...
    }
  }